State Management
& collision listening

+++ blox/src/com/blox/framework/v0/impl/libgdx/GdxTextureSplitter.java
@@ -14,21 +14,18 @@ class GdxTextureSplitter implements ITextureSplitter {
 	public ITexture[] split(ITexture texture, int width, int height) {
 		Texture gdxTexture = ((GdxTexture) texture).texture;
 
-		int cols = (int) (gdxTexture.getWidth() / width);
-		int rows = (int) (gdxTexture.getHeight() / height);
+		int cols = gdxTexture.getWidth() / width;
+		int rows = gdxTexture.getHeight() / height;
 
 		if (cols < 1 || rows < 1)
-			return new GdxTextureRegion[] { new GdxTextureRegion(
-					new TextureRegion(gdxTexture)) };
+			return new GdxTextureRegion[] { new GdxTextureRegion(new TextureRegion(gdxTexture)) };
 
-		TextureRegion[][] regions = TextureRegion.split(gdxTexture, width,
-				height);
+		TextureRegion[][] regions = TextureRegion.split(gdxTexture, width, height);
 
 		return toArray(cols, rows, regions);
 	}
 
-	private GdxTextureRegion[] toArray(int cols, int rows,
-			TextureRegion[][] regions) {
+	private GdxTextureRegion[] toArray(int cols, int rows, TextureRegion[][] regions) {
 		GdxTextureRegion[] gdxRegions = new GdxTextureRegion[cols * rows];
 		int index = 0;
 		for (int i = 0; i < rows; i++) {
